backend/requirements.txt
asgiref==3.7.1
django==3.2.19
djangorestframework==3.14.0
django-cors-headers==4.0.0
django-colorfield==0.8.0
django-extra-fields==3.0.2
django-filter==23.2
djoser==2.1.0
gunicorn==20.1.0
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dotenv==0.21.1
pytz==2023.3
reportlab==4.0.4
sqlparse==0.4.4
typing-extensions==4.6.1